+++ src/test/java/nl/jqno/equalsverifier/integration/extra_features/nonnull/jsr305/parametersarenonnullbydefault/package-info.java
@@ -1,0 +1,5 @@
+/** Applies JSR305's ParametersAreNonnullByDefault annotation to the whole package. */
+@ParametersAreNonnullByDefault
+package nl.jqno.equalsverifier.integration.extra_features.nonnull.jsr305.parametersarenonnullbydefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;