+++ CHANGELOG.md
@@ -13,9 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Initial support for Java 14's records and their [new equality invariant](https://docs.oracle.com/en/java/javase/14/docs/api/java.base/java/lang/Record.html).
+- Support for JSR305's `ParametersAreNonnullByDefault` annotation. ([Issue 308](https://github.com/jqno/equalsverifier/issues/308))
 ### Changed
 - Renames the project's main branch to `main`.
 - Replaces all references to the word 'black' to 'blue', including those in internal (but accessible) API's.
+### Fixed
 
 ## [3.3] - 2020-05-14
 ### Added